Add PSX BIOS from PSP 6.60
It has been optimised/enhanced/improved by Sony compared to the "traditional" PSX BIOS versions extracted from PSX hardware that we have all been using for years. This BIOS enables enhanced performance and compatibility in all emulators and is region free. It can be renamed.
Thanks to @thwqwer to have informed me of the existence of this BIOS